backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
